requirements.txt
# ============================================
# Pasaporte Médico Digital para Mascotas
# Dependencias del proyecto
# ============================================
# Para instalar: pip install -r requirements.txt

# Framework web principal
Flask==3.1.0

# Conector MySQL para Python (puro Python, no requiere compilar)
PyMySQL==1.1.1

# Para cargar variables de entorno desde un archivo .env
python-dotenv==1.1.0
